+++ CHANGELOG.md
@@ -22,7 +22,9 @@ cherry-pick it from the stable branch into master.
 * Added
   * None
 * Fixed
-  * None
+  * replace use of 1) <attribute>_changed? with will_save_change_to_<attribute>?,
+    2) <model>.changed? with <model>.has_changes_to_save?, and 3) <attribute>_was?
+    with <attribute>_in_database
 * Dependencies
   * Drop support for rails < 5.2
   * Add support for rails 6.0